fix: widen story_points to float, fix JIRA adapter int coercion (#56)
JIRA supports fractional story points (e.g., modified Fibonacci scale
0.5, 1.5, 3.5), but the JIRA adapter cast values to int(), silently
truncating 3.5 to 3 and corrupting velocity reports for teams using
those scales.

- UnifiedIssue.story_points and UnifiedSprint.{planned,completed}_story_points
  widened from Optional[int] to Optional[float]
- jira_adapter_converters._extract_story_points returns float; replaces
  int(float(value)) with float(value)
- pm_framework.base._extract_story_points base implementation also
  returns float, mirroring the JIRA adapter
- SQLAlchemy columns flipped from Integer to Float on cached_commits,
  qualitative_commits, daily_metrics, weekly_trends.story_points_delivered,
  pull_request_cache, issue_cache; SQLite jira_tickets schema
  uses REAL
- _DayStats.story_points dataclass field widened to float
- pipeline_report.py no longer truncates with int()
- New v12.0 migration helper logs the schema bump (SQLite dynamic
  type affinity means no DDL rebuild is needed)
- Adds tests/integrations/test_jira_adapter_story_points.py covering
  3.5, 1.5 (str), 0.5, integer-as-float, and unparseable-string cases
- Updates test_weekly_trends_velocity_columns to expect Float for
  story_points_delivered

Bumps version to 3.15.0.

+++ src/gitflow_analytics/core/metrics_storage.py
@@ -42,7 +42,9 @@ class _DayStats(TypedDict):
     files_changed: int
     lines_added: int
     lines_deleted: int
-    story_points: int
+    # WHY float (issue #56): aggregated story points may be fractional when
+    # underlying issues use modified Fibonacci scales (e.g., 0.5, 1.5, 3.5).
+    story_points: float
     tracked_commits: int
     untracked_commits: int
     unique_tickets: set[str]
@@ -68,7 +70,7 @@ def _make_day_stats() -> _DayStats:
         files_changed=0,
         lines_added=0,
         lines_deleted=0,
-        story_points=0,
+        story_points=0.0,
         tracked_commits=0,
         untracked_commits=0,
         unique_tickets=set(),
